More Multi-Architecture Support

+++ kernel/mem/paging.cpp
@@ -135,19 +135,19 @@ static void px_paging_map_hh_kernel() {
     }
 }
 
-static inline void px_set_page_dir(uint32_t page_dir) {
+static inline void px_set_page_dir(size_t page_dir) {
     asm volatile("mov %0, %%cr3" :: "b"(page_dir));
 }
 
 static inline void px_paging_enable() {
-    uint32_t cr0;
+    size_t cr0;
     asm volatile("mov %%cr0, %0": "=b"(cr0));
     cr0 |= 0x80000000;
     asm volatile("mov %0, %%cr0":: "b"(cr0));
 }
 
 static inline void px_paging_disable() {
-    uint32_t cr0;
+    size_t cr0;
     asm volatile("mov %%cr0, %0": "=b"(cr0));
     cr0 &= ~(0x80000000U);
     asm volatile("mov %0, %%cr0":: "b"(cr0));
